fastapi-demo-api/requirements.txt
FastAPI
Uvicorn
SQLAlchemy
Passlib[bcrypt]
python-jose
pydantic
email-validator
python-dotenv
pytest
httpx